renamed DataList to ListComponent
