--- src/components/DataList/index.ts
@@ -1,3 +1,0 @@
-import DataList from './data-list'
-
-export default DataList
