+++ src/components/ListComponent/index.ts
@@ -1,0 +1,3 @@
+import ListComponent from './list-component'
+
+export default ListComponent
