+++ src/components/index.ts
@@ -27,7 +27,7 @@ import IconButton from './IconButton'
 import ImageRadio from './ImageRadio'
 import InlineStepper from './InlineStepper'
 import Link from './Link'
-import ListComponent from './DataList'
+import ListComponent from './ListComponent'
 import Menu from './Menu'
 import NavigationLink from './NavigationLink'
 import Paginator from './Paginator'
